+++ CHANGELOG.md
@@ -1,17 +1,32 @@
+## HK Sounds release 4.0.0
+
+### Changes
+
+- Added option to enable/disable Killing Blow sounds.
+- Added Friendly Death sound support in Arenas (community requested feature).
+
+### Changes
+
+- Increased volume of the 'Gunshot 2' sound.
+- Added new sound 'Gunshot 3 (CS 1.6 AWP)'
+
 ## HK Sounds release 3.1.1
 
 ### Changes
+
 - Increased volume of the 'Gunshot 2' sound.
 - Added new sound 'Gunshot 3 (CS 1.6 AWP)'
 
 ## HK Sounds release 3.1.0
 
 ### Changes
+
 - It is possible to select multiple Single Sounds to play after a killing blow randomly.
 - A couple of new sounds added for Single Sound mode
 
 ## HK Sounds release 3.0.1
 
 ### Changes
+
 - GLaDOS sound pack added
 - Wilhelm Scream sound added for Single Sound mode